Enhanced account creation process
Added functionality to create a new account when a customer is created. The new account gets assigned with a random account number and default values for type and branch address. This change simplifies the process of creating an account for a new customer by automating it within the same service method.